[Feature]:Elapsed days,amenities,flat facilities
